generalized the CopyToClipboard component, separated out CopyAgentId component, and updated tests
