+++ __MOCKS__/copy-to-clipboard.js
