+++ src/holofuel/components/CopyToClipboard/__mocks__/copy-to-clipboard.js
@@ -1,0 +1,2 @@
+const copy = jest.fn()
+export default copy
